Manually merge latest backend changes.
